refactor(hir): split the ui-widget lowering test under the size cap
